* The legacy's GetFormat function has been re-implemented, but returns a   different Format struct than old versions of the library.
* frame_offset is now stored as an off64_t (for those oh-so-common cases when
  FRAMEOFFSET is > 2+9)

* dirfile_flush() man page added, and dirfile_close() man page corrected.

+++ src/open.c
@@ -524,7 +524,7 @@ static void _GD_ParseFormatFile(FILE* fp, DIRFILE *D, const char* filedir,
       /* Directives */
 
     } else if (strcmp(ptr, "FRAMEOFFSET") == 0) {
-      D->frame_offset = atoi(in_cols[1]);
+      D->frame_offset = atoll(in_cols[1]);
     } else if (strcmp(ptr, "INCLUDE") == 0) {
       int i, found = 0;
       char temp_buffer[FILENAME_MAX];
